refactor(phase10): migrate Constants to Kotlin object
- all String/Int/Boolean constants become const val (compiled static final with
  ConstantValue): AppConfigReceiver value refs and QuickSettingsTileForce's Java
  switch-case labels keep compiling unchanged
- PowerSource becomes a nested object with const vals (access syntax unchanged)
- DEFAULT_WEBGUI_TCP_PORT / DEFAULT_DATA_TCP_PORT unbox Integer -> Int;
  ApiClient.forceLoopbackHost's stale boxed-Integer comment + .toInt() removed
- REST_UPDATE_INTERVAL / GUI_UPDATE_INTERVAL stay runtime vals (SDK-dependent init)
- file helper functions stay object funs without @JvmStatic: all current Java files
  only reference constants, functions are called from Kotlin only

+++ app/src/main/java/com/nutomic/syncthingandroid/http/ApiClient.kt
@@ -264,8 +264,7 @@ class ApiClient(
  * would break the trust model above.
  */
 internal fun forceLoopbackHost(url: URL): URL {
-    // Constants.DEFAULT_WEBGUI_TCP_PORT is a boxed Integer on the Java side.
-    val port = if (url.port != -1) url.port else Constants.DEFAULT_WEBGUI_TCP_PORT.toInt()
+    val port = if (url.port != -1) url.port else Constants.DEFAULT_WEBGUI_TCP_PORT
     return try {
         URL(url.protocol, LOOPBACK_HOST, port, url.file)
     } catch (e: MalformedURLException) {
